backend/requirements.txt
aniso8601
attrs
brotlipy
cached-property
certifi
cffi
chardet
click
cryptography
ecdsa
Flask
Flask-Cors
Flask-JWT
Flask-JWT-Extended
Flask-RESTful
flask-restx
Flask-SQLAlchemy
idna
importlib-metadata
itsdangerous
Jinja2
jsonschema
python-keycloak==0.25.0
MarkupSafe
psycopg2
pyasn1
pycparser
PyJWT
pyOpenSSL
pyrsistent
PySocks
python-dateutil
python-jose
python-keycloak
pytz
PyYAML
requests==2.25.1
rsa
six
SQLAlchemy
urllib3==1.26.5
Werkzeug
zipp
neo4j==4.1.0
py2neo==2020.1.1
python-json-logger
pandas